+++ glue/src/test/java/org/somda/sdc/glue/consumer/sco/helper/OperationInvocationDispatcherTest.java
@@ -1,5 +1,6 @@
 package org.somda.sdc.glue.consumer.sco.helper;
 
+import com.google.inject.Injector;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -7,17 +8,20 @@ import org.somda.sdc.biceps.model.message.AbstractSetResponse;
 import org.somda.sdc.biceps.model.message.InvocationState;
 import org.somda.sdc.biceps.model.message.OperationInvokedReport;
 import org.somda.sdc.biceps.model.message.SetValueResponse;
+import org.somda.sdc.common.guice.AbstractConfigurationModule;
 import org.somda.sdc.dpws.ThisDeviceBuilder;
 import org.somda.sdc.dpws.ThisModelBuilder;
 import org.somda.sdc.dpws.service.HostingServiceProxy;
 import org.somda.sdc.dpws.service.factory.HostingServiceFactory;
 import org.somda.sdc.dpws.soap.RequestResponseClient;
 import org.somda.sdc.glue.UnitTestUtil;
+import org.somda.sdc.glue.consumer.ConsumerConfig;
 import org.somda.sdc.glue.consumer.sco.ScoArtifacts;
 import org.somda.sdc.glue.consumer.sco.ScoTransactionImpl;
 import org.somda.sdc.glue.consumer.sco.factory.OperationInvocationDispatcherFactory;
 import test.org.somda.common.LoggingTestWatcher;
 
+import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -27,12 +31,25 @@ import static org.mockito.Mockito.*;
 @ExtendWith(LoggingTestWatcher.class)
 class OperationInvocationDispatcherTest {
     private static final UnitTestUtil UT = new UnitTestUtil();
+
+    private Injector injector;
     private HostingServiceProxy hostingServiceProxy;
     private OperationInvocationDispatcher dispatcher;
 
+    private Duration timeout = Duration.ofMillis(500);
+
     @BeforeEach
     void beforeEach() {
-        hostingServiceProxy = UT.getInjector().getInstance(HostingServiceFactory.class).createHostingServiceProxy(
+        injector = UT.createInjectorWithOverrides(new AbstractConfigurationModule() {
+            @Override
+            protected void defaultConfigure() {
+                bind(ConsumerConfig.AWAITING_TRANSACTION_TIMEOUT,
+                    Duration.class,
+                    timeout);
+            }
+        });
+
+        hostingServiceProxy = injector.getInstance(HostingServiceFactory.class).createHostingServiceProxy(
                 "urn:uuid:441dfbea-40e5-406e-b2c4-154d3b8430bf",
                 Collections.emptyList(),
                 UT.getInjector().getInstance(ThisDeviceBuilder.class).get(),
@@ -42,7 +59,7 @@ class OperationInvocationDispatcherTest {
                 mock(RequestResponseClient.class),
                 "http://xAddr/");
 
-        dispatcher = UT.getInjector()
+        dispatcher = injector
                 .getInstance(OperationInvocationDispatcherFactory.class)
                 .createOperationInvocationDispatcher(hostingServiceProxy);
     }
@@ -98,5 +115,23 @@ class OperationInvocationDispatcherTest {
                     .receiveIncomingReport(report.getReportPart().get(i));
             verify(scoTransaction, atLeast(1)).getTransactionId();
         }
+    }
+
+    @Test
+    void concurrentDispatch() throws Exception {
+        final List<ScoTransactionImpl<? extends AbstractSetResponse>> transactions = new ArrayList<>();
+        long maxReports = 2;
+        OperationInvokedReport report = new OperationInvokedReport();
+        for (long i = 0; i < maxReports; ++i) {
+            ScoTransactionImpl<SetValueResponse> transaction = (ScoTransactionImpl<SetValueResponse>) mock(ScoTransactionImpl.class);
+            when(transaction.getTransactionId()).thenReturn(i);
+            transactions.add(transaction);
+            report.getReportPart().add(ScoArtifacts.createReportPart(i, InvocationState.FIN));
+            dispatcher.registerTransaction(transaction);
+        }
+
+        dispatcher.dispatchReport(report);
+        Thread.sleep(timeout.toMillis() * 2);
+        dispatcher.dispatchReport(report);
     }
 }